microservices/user-service/requirements.txt
grpcio
grpcio-tools
SQLAlchemy
psycopg2-binary